fix: UI fixes for widgets & Code structure
